localSync returns attributes like Backbone.sync, not a model
Fixes #32

localsync should be returning attributes, not a model, because it should
emulate how Backbone.sync acts. Unlike $.ajax, window.Store must receive
and return the model, not its attributes, since it must be able to
modify the correct idAttribute as specified by the model. Therefore it
must be localsync's responsibility to change the return value.