Force localsync callers to pass real Backbone models.
As a follow-up on issue / pr #51, this change forces localsync to be
called with backbone model or collection objects only.

First, a guard to `localsync` was added, similar as proposed by nilbus
[here](https://github.com/nilbus/Backbone.dualStorage/pull/51#issuecomment-25201923).
The model type restriction is not enfored for the localsync specific
extra methods (clear, hasDirtyOrDestroyed), since they don't use the
parameter at all. The guard was developed test-first to ensure correct
behavior.

Further on, all calls to localsync from dualstorage have been revisited.
The 'read', 'create' and 'update' methods now update the model through
its `set` method before the model is passed on to localsync (instead of
the received object). The changes have been verified through tests.