Consistent merging of local models with remote response.
The `model.set` operations on models prior passing them to localsync
have been refactored to a new helper function

    mergeModelWithRemoteResponse

which updates the model with the remote object properties and returns it
back to the caller.

To ensure that during update no 'change' events are triggered by
backbone, the `silent:true` option is passed along the set call.

All refactorings were accompanied with tests.