Update localsync with server responses on save/update
Fixes #55

Don't update update the persisted model's attributes; doing this stops
events from being emitted. Instead, create a clone to pass to localsync.
Using this clone allows localsync to know the model class's idAttribute
and store the update attributes from the server, without interfering
with the model being saved.

dualsync read should not merge attributes in this manner though.
When fetch is called, we want to pull the data from the server and pass
it as-is as the response to both localstorage and dualstorage.

The case we're trying to cover here, merging in dualsync, is updating
localsync on model save (create/update) when the server response is not
exactly what we sent in the save request.  The server response will get
processed by Backbone.Model.save and the model attributes will get
updated with the server response. Our localsync store gets left out of
the loop though, which is why we do this.

Test that dualsync fetch/read on a collection calls localsync on each
model in the collection in the response.

Don't test updateModelWithResponse/modelUpdatedWithResponse, because of
its role as a private method for dualsync rather than an external
collaborator. Instead, test for its effect on the public methods.

Rename serverReturnedAttributes to serverResponse, since the server can
return responses other than just one set of attributes.

When (read) fetching a collection, use a better local variable name
(collection) rather than model, which can be confusing.