Fix where fetching models/collections would not merge but overwrite locally stored attributes
This is mostly only an issue if the server does not respond with a full set of attributes.

modelUpdatedWithResponse is now always called with an existing model when possible,
We pass arguments in the constructor instead of no arguments to help with
https://github.com/nilbus/Backbone.dualStorage/issues/75#issuecomment-31745856,
thouh this isn't technically nessary.

Related to #37 and #55.