Fix syncDestroyed with custom idAttribute and backbone >= 1.1.1
syncDestroyed did not set the idAttribute attribute and only set
model.id. This worked in earlier versions of backbone, but it does not
as of 1.1.1. isNew now checks for @has(@idAttribute) instead of
model.id, as of
https://github.com/jashkenas/backbone/commit/11cc0e863f9951b4242144d65b581348bd9ee092.

Also, to avoid idAttribute issues in the future, stop setting model.id
anywhere. Instead, just set the idAttribute attribute, and let backbone
set model.id.