Prevent useOfflineStorage() from executing callbacks twice
This branch was allowing localsync to execute callbacks (ignoreCallbacks = false),
and unlike master, also explicitly calling success().

The the new variable in the spec is kept for clarity, but it was also causing
weirdness when the callback was being called twice: the 2nd time, the controller
variable was the DisconnectedCollection by the time the 2nd callback happened, and
assertions failed.